extract -> transcribe
https://github.com/wagtail/wagtail/issues/542#issuecomment-385798673

"Transcription is the term used by librarians & digital humanities researchers
for a plain text version of a document, either a photography, a video or an
audio document.

In lots of cases in digital humanities, we want to manually write
transcriptions instead of using OCR or extracting the already OCRed text from a
PDF. For example, medievalists transcribe documents almost impossible to OCR,
even today. Or musicians transcribe scores using languages such as LilyPond,
again almost impossible to OCR.

That’s why I think it’s better to have an editable transcription field. And for
consistency, use the verb transcribe for the functions/commands that
automatically fill the transcription. The transcription method itself should be
configurable, of course, so we can specify the backend and its options, like
mention we want Tesseract with these letters only and this dictionary, etc."